+++ mobile/android/android-components/components/browser/engine-gecko/src/test/java/mozilla/components/browser/engine/gecko/GeckoEngineSessionTest.kt
@@ -12,6 +12,8 @@ import android.view.WindowManager
 import androidx.test.ext.junit.runners.AndroidJUnit4
 import kotlinx.coroutines.ExperimentalCoroutinesApi
 import kotlinx.coroutines.test.runBlockingTest
+import mozilla.components.browser.engine.gecko.ext.geckoTrackingProtectionPermission
+import mozilla.components.browser.engine.gecko.ext.isExcludedForTrackingProtection
 import mozilla.components.browser.engine.gecko.permission.geckoContentPermission
 import mozilla.components.browser.errorpages.ErrorType
 import mozilla.components.concept.engine.DefaultSettings
@@ -68,7 +70,6 @@ import org.mockito.Mockito.verify
 import org.mockito.Mockito.verifyNoInteractions
 import org.mozilla.geckoview.AllowOrDeny
 import org.mozilla.geckoview.ContentBlocking
-import org.mozilla.geckoview.ContentBlockingController
 import org.mozilla.geckoview.GeckoResult
 import org.mozilla.geckoview.GeckoRuntime
 import org.mozilla.geckoview.GeckoSession
@@ -76,6 +77,10 @@ import org.mozilla.geckoview.GeckoSession.ContentDelegate.ContextElement.TYPE_AU
 import org.mozilla.geckoview.GeckoSession.ContentDelegate.ContextElement.TYPE_IMAGE
 import org.mozilla.geckoview.GeckoSession.ContentDelegate.ContextElement.TYPE_NONE
 import org.mozilla.geckoview.GeckoSession.ContentDelegate.ContextElement.TYPE_VIDEO
+import org.mozilla.geckoview.GeckoSession.PermissionDelegate.ContentPermission.VALUE_ALLOW
+import org.mozilla.geckoview.GeckoSession.PermissionDelegate.ContentPermission.VALUE_DENY
+import org.mozilla.geckoview.GeckoSession.PermissionDelegate.PERMISSION_STORAGE_ACCESS
+import org.mozilla.geckoview.GeckoSession.PermissionDelegate.PERMISSION_TRACKING
 import org.mozilla.geckoview.GeckoSession.ProgressDelegate.SecurityInformation
 import org.mozilla.geckoview.GeckoSessionSettings
 import org.mozilla.geckoview.SessionFinder
@@ -159,34 +164,52 @@ class GeckoEngineSessionTest {
     }
 
     @Test
-    @Suppress("DEPRECATION")
-    // The deprecation will be addressed on
-    // https://github.com/mozilla-mobile/android-components/issues/11101
     fun isIgnoredForTrackingProtection() {
-        val mockedContentBlockingController = mock<ContentBlockingController>()
-        var geckoResult = GeckoResult<Boolean?>()
         val session = GeckoEngineSession(runtime, geckoSessionProvider = geckoSessionProvider)
-        var wasExecuted = false
 
-        whenever(runtime.contentBlockingController).thenReturn(mockedContentBlockingController)
-        whenever(mockedContentBlockingController.checkException(any())).thenReturn(geckoResult)
+        session.geckoPermissions =
+            listOf(geckoContentPermission(type = PERMISSION_TRACKING, value = VALUE_ALLOW))
 
-        session.isIgnoredForTrackingProtection {
-            wasExecuted = it
-        }
+        var ignored = session.isIgnoredForTrackingProtection()
 
-        geckoResult.complete(true)
-        assertTrue(wasExecuted)
+        assertTrue(ignored)
 
-        geckoResult = GeckoResult()
-        whenever(mockedContentBlockingController.checkException(any())).thenReturn(geckoResult)
+        session.geckoPermissions =
+            listOf(geckoContentPermission(type = PERMISSION_TRACKING, value = VALUE_DENY))
 
-        session.isIgnoredForTrackingProtection {
-            wasExecuted = it
-        }
+        ignored = session.isIgnoredForTrackingProtection()
 
-        geckoResult.complete(null)
-        assertFalse(wasExecuted)
+        assertFalse(ignored)
+    }
+
+    @Test
+    fun `WHEN calling isExcludedForTrackingProtection THEN indicate if it is excluded for tracking protection`() {
+        val excludedPermission = geckoContentPermission(type = PERMISSION_TRACKING, value = VALUE_ALLOW)
+
+        assertTrue(excludedPermission.isExcludedForTrackingProtection)
+
+        val noExcludedPermission = geckoContentPermission(type = PERMISSION_TRACKING, value = VALUE_DENY)
+
+        assertFalse(noExcludedPermission.isExcludedForTrackingProtection)
+
+        val storagePermission = geckoContentPermission(type = PERMISSION_STORAGE_ACCESS, value = VALUE_DENY)
+
+        assertFalse(storagePermission.isExcludedForTrackingProtection)
+    }
+
+    @Test
+    fun `WHEN calling geckoTrackingProtectionPermission on a session THEN provide the gecko tracking protection permission`() {
+        val session = GeckoEngineSession(runtime, geckoSessionProvider = geckoSessionProvider)
+        val trackingProtectionPermission = geckoContentPermission(type = PERMISSION_TRACKING, value = VALUE_ALLOW)
+        val storagePermission = geckoContentPermission(type = PERMISSION_STORAGE_ACCESS, value = VALUE_DENY)
+
+        session.geckoPermissions = listOf(trackingProtectionPermission, storagePermission)
+
+        assertEquals(session.geckoTrackingProtectionPermission, trackingProtectionPermission)
+
+        session.geckoPermissions = listOf(storagePermission)
+
+        assertNull(session.geckoTrackingProtectionPermission)
     }
 
     @Test
@@ -240,12 +263,7 @@ class GeckoEngineSessionTest {
     }
 
     @Test
-    @Suppress("DEPRECATION")
-    // The deprecation will be addressed on
-    // https://github.com/mozilla-mobile/android-components/issues/11101
     fun navigationDelegateNotifiesObservers() {
-        val geckoResult = GeckoResult<Boolean?>()
-        val mockedContentBlockingController = mock<ContentBlockingController>()
         val engineSession = GeckoEngineSession(runtime, geckoSessionProvider = geckoSessionProvider)
 
         var observedUrl = ""
@@ -259,16 +277,10 @@ class GeckoEngineSessionTest {
             }
         })
 
-        whenever(runtime.contentBlockingController).thenReturn(mockedContentBlockingController)
-        whenever(mockedContentBlockingController.checkException(any())).thenReturn(geckoResult)
-
         captureDelegates()
-
-        geckoResult.complete(true)
 
         navigationDelegate.value.onLocationChange(mock(), "http://mozilla.org", emptyList())
         assertEquals("http://mozilla.org", observedUrl)
-        verify(mockedContentBlockingController).checkException(any())
 
         navigationDelegate.value.onCanGoBack(mock(), true)
         assertEquals(true, observedCanGoBack)
@@ -277,8 +289,6 @@ class GeckoEngineSessionTest {
         assertEquals(true, observedCanGoForward)
     }
 
-    // This will be addressed on https://github.com/mozilla-mobile/android-components/issues/8312
-    @Suppress("deprecation")
     @Test
     fun contentDelegateNotifiesObserverAboutDownloads() {
         val engineSession = GeckoEngineSession(
@@ -657,12 +667,7 @@ class GeckoEngineSessionTest {
     }
 
     @Test
-    @Suppress("DEPRECATION")
-    // The deprecation will be addressed on
-    // https://github.com/mozilla-mobile/android-components/issues/11101
     fun navigationDelegateIgnoresInitialLoadOfAboutBlank() {
-        val geckoResult = GeckoResult<Boolean?>()
-        val mockedContentBlockingController = mock<ContentBlockingController>()
         val engineSession = GeckoEngineSession(runtime, geckoSessionProvider = geckoSessionProvider)
 
         var observedUrl = ""
@@ -670,12 +675,7 @@ class GeckoEngineSessionTest {
             override fun onLocationChange(url: String) { observedUrl = url }
         })
 
-        whenever(runtime.contentBlockingController).thenReturn(mockedContentBlockingController)
-        whenever(mockedContentBlockingController.checkException(any())).thenReturn(geckoResult)
-
         captureDelegates()
-
-        geckoResult.complete(true)
 
         navigationDelegate.value.onLocationChange(mock(), "about:blank", emptyList())
         assertEquals("", observedUrl)
@@ -685,22 +685,13 @@ class GeckoEngineSessionTest {
 
         navigationDelegate.value.onLocationChange(mock(), "https://www.mozilla.org", emptyList())
         assertEquals("https://www.mozilla.org", observedUrl)
-        verify(mockedContentBlockingController).checkException(any())
 
         navigationDelegate.value.onLocationChange(mock(), "about:blank", emptyList())
         assertEquals("about:blank", observedUrl)
     }
 
     @Test
-    @Suppress("DEPRECATION")
-    // The deprecation will be addressed on
-    // https://github.com/mozilla-mobile/android-components/issues/11101
     fun `onLoadRequest will reset initial load flag on process switch to ignore about blank loads`() {
-        val geckoResult = GeckoResult<Boolean?>()
-        val mockedContentBlockingController = mock<ContentBlockingController>()
-        whenever(runtime.contentBlockingController).thenReturn(mockedContentBlockingController)
-        whenever(mockedContentBlockingController.checkException(any())).thenReturn(geckoResult)
-
         val session = GeckoEngineSession(runtime, geckoSessionProvider = geckoSessionProvider)
         captureDelegates()
         assertTrue(session.initialLoad)
@@ -743,15 +734,9 @@ class GeckoEngineSessionTest {
     }
 
     @Test
-    @Suppress("DEPRECATION")
-    // The deprecation will be addressed on
-    // https://github.com/mozilla-mobile/android-components/issues/11101
     fun `keeps track of current url via onLocationChange events`() {
-        val mockedContentBlockingController = mock<ContentBlockingController>()
         val engineSession = GeckoEngineSession(runtime, geckoSessionProvider = geckoSessionProvider)
         val geckoResult = GeckoResult<Boolean?>()
-        whenever(runtime.contentBlockingController).thenReturn(mockedContentBlockingController)
-        whenever(mockedContentBlockingController.checkException(any())).thenReturn(geckoResult)
 
         captureDelegates()
         geckoResult.complete(true)
@@ -765,22 +750,36 @@ class GeckoEngineSessionTest {
     }
 
     @Test
-    @Suppress("DEPRECATION")
-    // The deprecation will be addressed on
-    // https://github.com/mozilla-mobile/android-components/issues/11101
+    fun `WHEN onLocationChange is called THEN geckoPermissions is assigned`() {
+        val engineSession = GeckoEngineSession(runtime, geckoSessionProvider = geckoSessionProvider)
+
+        captureDelegates()
+
+        navigationDelegate.value.onLocationChange(geckoSession, "https://www.mozilla.org", listOf(mock()))
+
+        assertTrue(engineSession.geckoPermissions.isNotEmpty())
+    }
+
+    @Test
+    fun `WHEN onLocationChange is called with null URL THEN geckoPermissions is assigned`() {
+        val engineSession = GeckoEngineSession(runtime, geckoSessionProvider = geckoSessionProvider)
+
+        captureDelegates()
+
+        navigationDelegate.value.onLocationChange(geckoSession, null, listOf(mock()))
+
+        assertTrue(engineSession.geckoPermissions.isNotEmpty())
+    }
+
+    @Test
     fun `notifies configured history delegate of title changes`() = runBlockingTest {
-        val mockedContentBlockingController = mock<ContentBlockingController>()
         val engineSession = GeckoEngineSession(
             runtime, geckoSessionProvider = geckoSessionProvider,
             context = coroutineContext
         )
         val historyTrackingDelegate: HistoryTrackingDelegate = mock()
-        val geckoResult = GeckoResult<Boolean?>()
-        whenever(runtime.contentBlockingController).thenReturn(mockedContentBlockingController)
-        whenever(mockedContentBlockingController.checkException(any())).thenReturn(geckoResult)
 
         captureDelegates()
-        geckoResult.complete(true)
 
         // Nothing breaks if history delegate isn't configured.
         contentDelegate.value.onTitleChange(geckoSession, "Hello World!")
@@ -830,19 +829,13 @@ class GeckoEngineSessionTest {
     }
 
     @Test
-    @Suppress("DEPRECATION")
-    // The deprecation will be addressed on
-    // https://github.com/mozilla-mobile/android-components/issues/11101
     fun `notifies configured history delegate of preview image URL changes`() = runBlockingTest {
-        val mockedContentBlockingController = mock<ContentBlockingController>()
         val engineSession = GeckoEngineSession(
             runtime, geckoSessionProvider = geckoSessionProvider,
             context = coroutineContext
         )
         val historyTrackingDelegate: HistoryTrackingDelegate = mock()
         val geckoResult = GeckoResult<Boolean?>()
-        whenever(runtime.contentBlockingController).thenReturn(mockedContentBlockingController)
-        whenever(mockedContentBlockingController.checkException(any())).thenReturn(geckoResult)
 
         captureDelegates()
         geckoResult.complete(true)